[mathml] Add basic support tabular elements
This adds some basic CSS rules for <mtable>, <mtr> and <mtd>.

Apart from that it also includes a small change
in NGTableLayoutAlgorithm::GenerateFragment() in order to
vertically align the <mtable> with the math axis:
https://w3c.github.io/mathml-core/#table-or-matrix-mtable

This change adds a new test table-003, that doesn't pass right now.
The problem is that the <mtd> is not horizontally centering its contents.

Bug: 6606, 1125111
Test: external/wpt/mathml/presentation-markup/tables/table-001.html
Test: external/wpt/mathml/presentation-markup/tables/table-002.html
Test: external/wpt/mathml/presentation-markup/tables/table-003.html
Change-Id: I536f1e8344baf1ed4de431f2da00aa55dbe3aa6e