Don't unpack posted messages bound for remote windows.
Doing so causes them to lose their transferable data and later fail
deserialization. This happens implicitly when a MessageEvent is
constructed, so don't construct a MessageEvent.

Instead, use a simple struct to pass the needed data around and then
either unpack it into a MessageEvent (local) or convert it to Mojo IPC
data (remote) as appropriate.

Bug: 1212748
Change-Id: Iab433a23f9e7db1406ccbb14ee7ef73258cf2cad